# Optimize binary pattern matching in Integer.parse